+++ InputHandler.h
@@ -5,10 +5,10 @@
 class InputHandler{
 public:
     std::string readFileName();
-    void readSize(int &height, int &width);
+    void readSize(short &height, short &width);
     int readWord(Word &word);
 private:
-    int m_height, m_width;
+    short m_height, m_width;
     bool checkFileName(const std::string &name) const;
     bool checkWordInput(const std::string &input) const;
 };
